+++ main/isotp_user.h
@@ -6,7 +6,7 @@ void isotp_user_debug(const char* message, ...);
 
 /* user implemented, send can message */
 int  isotp_user_send_can(const uint32_t arbitration_id,
-                         const uint8_t* data, const uint8_t size);
+                         const uint8_t* data, const uint16_t size);
 
 /* user implemented, get microsecond */
 uint64_t isotp_user_get_us(void);
